cping-bot-telegram/requirements.txt
python-telegram-bot
requests
python-dotenv
supabase